fix: address PR review feedback
Sort com_ch_* i18n keys alphabetically in translation.json.
Revert PWA precache limit to 4MB (vendor is 3.4MB with chunk isolation).
Move formatCHC into CostView to keep cost helpers colocated.
Fix UTC date parsing in CostView with parseLocalDate helper.
Build FlatTable columns from all rows, not just first row.
Fix grid bottom padding gap.
