+++ client/src/locales/en/translation.json
@@ -219,6 +219,20 @@
   "com_citation_more_details": "More details about {{label}}",
   "com_citation_source": "Source",
   "com_click_to_download": "(click here to download)",
+  "com_ch_collapse_all": "Collapse all",
+  "com_ch_cost_by_date": "By Date",
+  "com_ch_cost_by_entity": "By Entity",
+  "com_ch_cost_daily": "Daily",
+  "com_ch_cost_total": "Total: {{value}}",
+  "com_ch_cost_weekly": "Weekly",
+  "com_ch_expand_all": "Expand all",
+  "com_ch_filter_columns": "Filter Columns ({{visible}}/{{total}})",
+  "com_ch_label_elapsed": "Elapsed: {{value}}",
+  "com_ch_label_read": "Read: {{rows}} rows",
+  "com_ch_label_read_bytes": "Read: {{rows}} rows ({{bytes}})",
+  "com_ch_label_row": "{{count}} row",
+  "com_ch_label_rows": "{{count}} rows",
+  "com_ch_tab_query": "Query",
   "com_download_expired": "(download expired)",
   "com_download_expires": "(click here to download - expires {{0}})",
   "com_endpoint": "Endpoint",
@@ -1581,19 +1595,5 @@
   "com_ui_zoom_in": "Zoom in",
   "com_ui_zoom_level": "Zoom level",
   "com_ui_zoom_out": "Zoom out",
-  "com_user_message": "You",
-  "com_ch_tab_query": "Query",
-  "com_ch_label_elapsed": "Elapsed: {{value}}",
-  "com_ch_label_read": "Read: {{rows}} rows",
-  "com_ch_label_read_bytes": "Read: {{rows}} rows ({{bytes}})",
-  "com_ch_label_row": "{{count}} row",
-  "com_ch_label_rows": "{{count}} rows",
-  "com_ch_expand_all": "Expand all",
-  "com_ch_collapse_all": "Collapse all",
-  "com_ch_filter_columns": "Filter Columns ({{visible}}/{{total}})",
-  "com_ch_cost_total": "Total: {{value}}",
-  "com_ch_cost_by_entity": "By Entity",
-  "com_ch_cost_by_date": "By Date",
-  "com_ch_cost_daily": "Daily",
-  "com_ch_cost_weekly": "Weekly"
+  "com_user_message": "You"
 }
